+++ src/app/app-routing.module.ts
@@ -2,14 +2,12 @@ import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { PartydirectoryComponent } from './partydirectory/partydirectory.component';
 import { WelcomeComponent } from './welcome/welcome.component';
-import { MypartiesComponent } from './myparties/myparties.component';
 import { MycharactersComponent } from './mycharacters/mycharacters.component';
 import { AuthGuard } from './auth/auth.guard';
 
 const routes: Routes = [
   { path: 'partydirectory', component: PartydirectoryComponent },
   { path: 'welcome', component: WelcomeComponent },
-  { path: 'myparties', component: MypartiesComponent, canActivate: [AuthGuard] },
   { path: 'mycharacters', component: MycharactersComponent, canActivate: [AuthGuard] },
   { path: '', redirectTo: '/welcome', pathMatch: 'full' }
 ];
