+++ src/app/dialog/confirm-dialog.html
@@ -5,6 +5,6 @@
   </p>
 </div>
 <div mat-dialog-actions align="end">
-  <button mat-stroked-button color="warn" (click)="onCancel">Cancel</button>
+  <button mat-stroked-button color="warn" (click)="onCancel()">Cancel</button>
   <button mat-stroked-button (click)="onOk()">OK</button>
 </div>
