+++ src/app/header/header.component.html
@@ -11,35 +11,7 @@
     <td>
       <img src="../assets/img/spacer.png" width="10px">
     </td>
-    <td class="header-search" style="width:80%;">
-    <!-- TODO: Functional implementation
-      <form>
-        <table>
-          <tr>
-            <td>
-              <div class="search-field">
-                <input type="text" name="s" placeholder="Search Tag(s)" size="60">
-              </div>
-            </td>
-            <td>
-              <div class="search-button-container">
-                <button class="search-button" title="Search" mat-icon-button>
-                  <mat-icon>search</mat-icon>
-                </button>
-              </div>
-            </td>
-            <td>
-              <div class="help-button-container">
-                <button class="help-button" (click)="toggleHelp()" title="Help" type="button" mat-icon-button>
-                  <mat-icon>help</mat-icon>
-                </button>
-              </div>
-            </td>
-          </tr>
-        </table>
-      </form>
-    --> 
-    </td>
+    <td style="width:80%"></td>
     <td id="header-controlpanel" style="width:20%">
       <app-controlpanel></app-controlpanel>
     </td>
