+++ src/app/header/header.component.ts
@@ -1,9 +1,6 @@
-import { Component, OnInit, Input, Inject } from '@angular/core';
-import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
-
-export interface HelpDialogData {
-  instance: string;
-}
+import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material';
+import { FormGroup, FormControl} from '@angular/forms';
 
 @Component({
   selector: 'app-header',
@@ -12,31 +9,7 @@ export interface HelpDialogData {
 })
 export class HeaderComponent implements OnInit {
 
-  constructor(public dialog: MatDialog) { }
-
-  toggleHelp(): void {
-    const dialogRef = this.dialog.open(HeaderHelpDialog,
-      {
-        autoFocus: false,
-        width: '90vw',
-        maxWidth: '700px',
-        maxHeight: '100vh',
-      });
-  }
+  constructor() {  }
 
   ngOnInit() { }
-}
-
-@Component({
-  selector: 'header-help-dialog',
-  templateUrl: 'header-help-dialog.html',
-})
-export class HeaderHelpDialog {
-  constructor(
-    public dialogRef: MatDialogRef<HeaderHelpDialog>,
-    @Inject(MAT_DIALOG_DATA) public data: HelpDialogData) { }
-
-  onOk() {
-    this.dialogRef.close();
-  }
 }
