--- src/app/myparties/myparties.component.html
@@ -1,24 +1,0 @@
-<div class="myparties-container">
-  <div class="secondary-nav">
-    <mat-progress-spinner *ngIf="isLoading && (!hasFetchedInstances || !hasFetchedCharacters || !hasFetchedPurposes || !hasFetchedJobs)" class="spinner" color="primary" mode="indeterminate" diameter="40" strokeWidth="7"></mat-progress-spinner>
-    <div class="progress-notifications">
-      <span *ngIf="!hasFetchedInstances && timeLeft > 0">Fetching Instance Data...</span>
-      <br *ngIf="!hasFetchedCharacters && timeLeft > 0"><span *ngIf="!hasFetchedCharacters && timeLeft > 0">Fetching Character Data...</span>
-      <br *ngIf="!hasFetchedJobs && timeLeft > 0"><span *ngIf="!hasFetchedJobs && timeLeft > 0">Fetching Job Data...</span>
-      <br *ngIf="!hasFetchedPurposes && timeLeft > 0"><span *ngIf="!hasFetchedPurposes && timeLeft > 0">Fetching Party Purpose Data...</span>
-      <br *ngIf="!hasFetchedParties && timeLeft > 0"><span *ngIf="!hasFetchedParties && timeLeft > 0">Fetching Parties...</span>
-      <br *ngIf="isLoading && timeLeft == 0"><span *ngIf="isLoading && timeLeft == 0">Failed to Retrieve Data.<br><br> <button mat-raised-button (click)="retryFetchData()">Try Again</button></span>
-    </div>
-    <button *ngIf="!isLoading && hasFetchedInstances && hasFetchedCharacters && hasFetchedPurposes && hasFetchedJobs" mat-raised-button class="create-button" (click)="confirmCreate()" title="Create a New Party">
-      <mat-icon>
-        create
-      </mat-icon>
-      New Party
-    </button>
-  </div>
-  <mat-paginator [length]="length" [pageSize]="pageSize" [pageSizeOptions]="pageSizeOptions" [pageIndex]="currentPage-1" (page)="onChangePage($event)"></mat-paginator>
-  <app-partycard [partyDetails]="party" [characters]="characters" *ngFor="let party of parties"></app-partycard>
-  <div class="empty-party-list" *ngIf="parties.length <= 0 && !isLoading && hasFetchedParties">You aren't currently in any parties, select "New Party"
-    above or visit the Directory to get started.</div>
-  <mat-paginator [length]="length" [pageSize]="pageSize" [pageSizeOptions]="pageSizeOptions" [pageIndex]="currentPage-1" (page)="onChangePage($event)"></mat-paginator>
-</div>
