+++ src/app/partydirectory/partydirectory.component.html
@@ -1,10 +1,29 @@
-<div class="spinner" *ngIf="isLoading">
-    <mat-progress-spinner class="spinner" color="primary" mode="indeterminate" diameter="40" strokeWidth="7"></mat-progress-spinner>
-    <div class="progress-notifications">
-      <span style="text-align:center" *ngIf="isLoading && timeLeft > 0">Fetching Parties...</span>
-      <br *ngIf="isLoading && timeLeft == 0"><span *ngIf="isLoading && timeLeft == 0">Failed to Retrieve Data.<br><br> <button mat-raised-button (click)="retryFetchData()">Try Again</button></span>
-    </div>
+<div class="spinner">
+  <mat-progress-spinner *ngIf="isLoading && (!hasFetchedInstances || !hasFetchedCharacters || !hasFetchedPurposes || !hasFetchedJobs)" class="spinner" color="primary" mode="indeterminate" diameter="40" strokeWidth="7"></mat-progress-spinner>
+  <div class="progress-notifications">
+    <span *ngIf="!hasFetchedInstances && timeLeft > 0">Fetching Instance Data...</span>
+    <br *ngIf="!hasFetchedCharacters && timeLeft > 0"><span *ngIf="!hasFetchedCharacters && timeLeft > 0">Fetching Character Data...</span>
+    <br *ngIf="!hasFetchedJobs && timeLeft > 0"><span *ngIf="!hasFetchedJobs && timeLeft > 0">Fetching Job Data...</span>
+    <br *ngIf="!hasFetchedPurposes && timeLeft > 0"><span *ngIf="!hasFetchedPurposes && timeLeft > 0">Fetching Party Purpose Data...</span>
+    <br *ngIf="!hasFetchedParties && timeLeft > 0"><span *ngIf="!hasFetchedParties && timeLeft > 0">Fetching Parties...</span>
+    <br *ngIf="isLoading && timeLeft == 0"><span *ngIf="isLoading && timeLeft == 0">Failed to Retrieve Data.<br><br> <button mat-raised-button (click)="retryFetchData()">Try Again</button></span>
   </div>
+  <button mat-raised-button class="create-button" *ngIf="isAuth && !isLoading && hasFetchedInstances && hasFetchedCharacters && hasFetchedPurposes && hasFetchedJobs" (click)="confirmCreate()" title="Create a New Party">
+    <mat-icon>
+      create
+    </mat-icon>
+    New Party
+  </button>
+  <button mat-raised-button class="filter-button" *ngIf="!isLoading && hasFetchedInstances && hasFetchedCharacters && hasFetchedPurposes && hasFetchedJobs" (click)="toggleFilters()" title="Show/Hide Filters">
+    <mat-icon>
+      filter_list
+    </mat-icon>
+    Toggle Filters
+  </button>
+</div>
+<div *ngIf="displayFilters" class="search-filters">
+  <app-partyfilter></app-partyfilter>
+</div>
 <div class="parties-container" *ngIf="!isLoading">
   <mat-paginator [length]="length" [pageSize]="pageSize" [pageSizeOptions]="pageSizeOptions" [pageIndex]="currentPage-1" (page)="onChangePage($event)"></mat-paginator>
   <app-partycard [partyDetails]="party" *ngFor="let party of parties"></app-partycard>
