+++ src/app/partydirectory/partydirectory.component.scss
@@ -1,11 +1,15 @@
 @import "/../app.component";
 
+.filter-button {
+  margin-left: 5px;
+}
+
 .spinner{
   text-align: center;
-  margin-top: 30px;
+  margin-top: 10px;
   margin-left: auto;
   margin-right: auto;
-  margin-bottom: 5px;
+  margin-bottom: 0px;
   background-color: transparent;
 }
 
@@ -18,7 +22,7 @@ div {
 }
 
 .progress-notifications {
-  margin-top: 20px;
+  margin-top: 0px;
   background-color: transparent;
 }
