+++ src/app/primarynav/partyfilter/partyfilter.component.html
@@ -1,0 +1,65 @@
+<div class="filters" [formGroup]="form">
+  <h2>Filters</h2>
+  <mat-form-field (change)="setInstance()">
+    <mat-label>Instance</mat-label>
+    <select formControlName="instance" matNativeControl>
+      <option *ngFor="let instance of this.instances" [value]="instance">{{ instance }}</option>
+    </select>
+  </mat-form-field>
+  <mat-form-field (change)="setPurpose()">
+    <mat-label>Purpose</mat-label>
+    <select formControlName="purpose" matNativeControl>
+      <option *ngFor="let purpose of this.purposes" [value]="purpose">{{ purpose }}</option>
+    </select>
+  </mat-form-field>
+  <mat-form-field (change)="setJob()">
+    <mat-label>Job Availability (Not Implemented)</mat-label>
+    <select formControlName="job" matNativeControl>
+      <option *ngFor="let job of this.jobs" [value]="job">{{ job }}</option>
+    </select>
+  </mat-form-field>
+  <mat-form-field (change)="setDifficulty()">
+    <mat-label>Difficulty (Not Implemented)</mat-label>
+    <select formControlName="difficulty" matNativeControl>
+      <option *ngFor="let difficulty of this.difficulties" [value]="difficulty">{{ difficulty }}</option>
+    </select>
+  </mat-form-field>
+  <mat-form-field (change)="setIType()">
+    <mat-label>Instance Type (Not Implemented)</mat-label>
+    <select formControlName="itype" matNativeControl>
+      <option *ngFor="let itype of this.itypes" [value]="itype">{{ itype }}</option>
+    </select>
+  </mat-form-field>
+  <!--<mat-form-field>
+    <mat-label>Minimum iLevel</mat-label>
+    <mat-slider min="0" max="500" step="5" value="0" matNativeControl formControlName="ilvl"></mat-slider>
+  </mat-form-field>
+  <mat-form-field>
+    <mat-label>Minimum Level</mat-label>
+    <mat-slider min="0" max="80" step="1" value="0" matNativeControl formControlName="lvl"></mat-slider>
+  </mat-form-field>
+  <mat-form-field>
+    <mat-slide-toggle matNativeControl formControlName="verf">
+      Verified Users Only
+    </mat-slide-toggle>
+  </mat-form-field>
+  <mat-form-field>
+    <mat-label>Minimum Parse</mat-label>
+    <mat-slider min="0" max="100" step="1" value="0" matNativeControl formControlName="parse"></mat-slider>
+  </mat-form-field>
+  <mat-form-field>
+    <mat-label>Minimum Parse Median</mat-label>
+    <mat-slider min="0" max="100" step="1" value="0" matNativeControl formControlName="parsem"></mat-slider>
+  </mat-form-field> -->
+  <!-- Owner Name TODO -->
+  <!-- Party Member Name TODO -->
+  <!-- Description Contains TODO -->
+  <!-- Schedule TODO -->
+  <br>
+  <button mat-raised-button class="filter-button" (click)="filterSearch()" title="Filtered Search">
+    <mat-icon class="filter-search-icon">
+      search
+    </mat-icon>
+    Filter Search
+  </button>
+</div>
