+++ src/app/primarynav/partyfilter/partyfilter.component.scss
@@ -1,0 +1,16 @@
+@import "../../app.component";
+
+div {
+  margin-top: -7px;
+}
+
+.filters {
+  font-family: Arial;
+  font-size: 0.875em;
+  text-align: center;
+}
+
+.filter-search-icon {
+  padding-right: 5px;
+  margin: none;
+}
