+++ src/app/primarynav/primarynav.component.html
@@ -1,10 +1,5 @@
 <div class="primary-nav">
   <mat-nav-list flex fxFlexAlign="center">
-    <button *ngIf="isAuthenticated()" mat-icon-button>
-      <mat-icon class="nav-button" title="My Parties" routerLink="/myparties" routerLinkActive="gold">
-        group
-      </mat-icon>
-    </button>
     <button mat-icon-button>
       <mat-icon class="nav-button" title="Directory" routerLink="/partydirectory" routerLinkActive="gold">
         list
