+++ src/app/primarynav/primarynav.component.scss
@@ -4,3 +4,7 @@
   margin-top: 10px;
   padding: 0px;
 }
+
+.search-filters {
+  outline: none;
+}
