+++ src/app/primarynav/primarynav.component.ts
@@ -2,11 +2,9 @@ import { Component, OnInit, Input, Inject } from '@angular/core';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { Subscription } from 'rxjs';
 import { AuthService } from '../auth/auth.service';
+import { FormBuilder, FormGroup } from '@angular/forms';
 
-export interface MessagesDialogData {
-  //instance: string;
-
-}
+export interface MessagesDialogData {}
 
 @Component({
   selector: 'app-primarynav',
@@ -29,12 +27,10 @@ export class PrimarynavComponent implements OnInit {
         width: '90vw',
         maxWidth: '600px',
         maxHeight: '90%',
-        data: {
-        }
       });
   }
 
-  constructor(public dialog: MatDialog, private as: AuthService) { }
+  constructor(public dialog: MatDialog, private as: AuthService) {}
 
   ngOnInit()
   {
